[WebGPU] Make wgslc and wgslfuzz distinct from each other https://bugs.webkit.org/show_bug.cgi?id=259454 rdar://112798797
Reviewed by Mike Wyrzykowski.

It would be useful to make the compiler and the fuzzer different executables
for local development, so when the fuzzer finds a problem, you can feed that
into the compiler to debug it.

The way this works is:
% set-webkit-configuration --debug --asan --libFuzzer
% cd Source/WebGPU
% make SCHEME=wgslfuzz
% make SCHEME=wgslc
% ASAN_OPTIONS=whatever DYLD_FRAMEWORK_PATH=/path/to/Products/Debug DYLD_LIBRARY_PATH=/path/to/Products/Debug /path/to/Products/Debug/wgslfuzz
% ASAN_OPTIONS=whatever DYLD_FRAMEWORK_PATH=/path/to/Products/Debug DYLD_LIBRARY_PATH=/path/to/Products/Debug /path/to/Products/Debug/wgslc

* Source/WebGPU/Configurations/wgslfuzz.xcconfig: Added.
* Source/WebGPU/WGSL/wgslc.cpp:
(main):
(LLVMFuzzerTestOneInput): Deleted.
* Source/WebGPU/WGSL/wgslfuzz.cpp: Added.
(LLVMFuzzerTestOneInput):
* Source/WebGPU/WebGPU.xcodeproj/project.pbxproj:
* Source/WebGPU/WebGPU.xcodeproj/xcshareddata/xcschemes/wgslfuzz.xcscheme: Added.

Canonical link: https://commits.webkit.org/266269@main
