+++ Source/WebGPU/WGSL/wgslfuzz.cpp
@@ -1,0 +1,53 @@
+/*
+ * Copyright (c) 2023 Apple Inc. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY APPLE INC. ``AS IS'' AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL APPLE INC. OR
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
+ * OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include "config.h"
+
+#include "AST/ASTStringDumper.h"
+#include "WGSL.h"
+#include "WGSLShaderModule.h"
+#include <wtf/DataLog.h>
+#include <wtf/FileSystem.h>
+#include <wtf/WTFProcess.h>
+
+extern "C" {
+int LLVMFuzzerTestOneInput(const uint8_t *data, size_t);
+int LLVMFuzzerInitialize(int *argc, char ***argv);
+} // extern "C"
+
+int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size)
+{
+    WTF::initializeMainThread();
+
+    WGSL::Configuration configuration;
+    auto source = String::fromUTF8WithLatin1Fallback(data, size);
+    auto checkResult = WGSL::staticCheck(source, std::nullopt, configuration);
+    if (auto* successfulCheck = std::get_if<WGSL::SuccessfulCheck>(&checkResult)) {
+        auto& shaderModule = successfulCheck->ast;
+        WGSL::prepare(shaderModule, "main"_str, std::nullopt);
+    }
+
+    return 0;
+}
